Initial Commit: Easi.work Simple Demo
This is a simple Email interface using custom domains and an SMTP/IMAP
server for a proof of concept. The goal is to be able to kick off agent swarms,
include human-in-the-loop solutions, and to interact directly with user
interfaces on behalf of people. This script runs on a cronjob managed
locally every minute.